requirements.txt
asgiref==3.6.0
certifi==2023.5.7
charset-normalizer==3.1.0
dj-database-url==0.5.0
Django==3.2.18
gunicorn==20.1.0
idna==3.4
numpy
pandas==1.3.5
Pillow==9.5.0
psycopg2-binary==2.9.6
python-dateutil==2.8.2
pytz==2023.3
requests==2.30.0
six==1.16.0
sqlparse==0.4.4
stripe==5.4.0
typing_extensions==4.5.0
urllib3==2.0.2
whitenoise==6.4.0
django-cloudinary-storage
django-jazzmin
pyfcm
django-twilio
psycopg2-binary
fcm-django
django-crontab
channels
xhtml2pdf
django-mathfilters
django_environ